restructure tests (#265)
* restructure tests

* improve tests
