+++ spec/features/chunked/chunked_reading_spec.rb
@@ -4,7 +4,7 @@ require 'spec_helper'
 
 fixture_path = 'spec/fixtures'
 
-describe 'be_able_to' do
+describe 'chunked reading' do
   it 'loads_chunk_cornercase_csv_files' do
     6.times do |chunk_size| # test for all chunk-sizes
       options = {chunk_size: chunk_size, remove_empty_hashes: true}
