+++ spec/features/converters/value_converters_spec.rb
@@ -17,7 +17,7 @@ class CurrencyConverter
   end
 end
 
-describe 'be_able_to' do
+describe ':value_converters option' do
   it 'convert date values into Date instances' do
     options = {value_converters: {date: DateConverter}}
     data = SmarterCSV.process("#{fixture_path}/with_dates.csv", options)
