+++ spec/features/formating/bom_issues_spec.rb
