+++ spec/features/formating/emoji_spec.rb
