+++ spec/features/formating/line_ending_spec.rb
