+++ spec/features/general/additional_separator_spec.rb
