+++ spec/features/general/empty_columns_spec.rb
