+++ spec/features/hash_transformations/remove_not_mapped_keys_spec.rb
