+++ spec/features/hash_transformations/remove_zero_values_spec.rb
@@ -4,8 +4,8 @@ require 'spec_helper'
 
 fixture_path = 'spec/fixtures'
 
-describe 'be_able_to' do
-  it 'remove_zero_values' do
+describe ':remove_zero_values option' do
+  it 'removes zero values' do
     options = {remove_zero_values: true, remove_empty_values: true}
     data = SmarterCSV.process("#{fixture_path}/basic.csv", options)
     expect(data.size).to eq 5
@@ -17,7 +17,7 @@ describe 'be_able_to' do
         expect(%i[first_name last_name dogs cats birds fish]).to include(key)
       end
 
-      expect(hash.values).to_not include(0)
+      expect(hash.values).not_to include(0)
 
       expect(hash.size).to be <= 6
     end
