+++ spec/features/header_handling/duplicate_headers_spec.rb
