+++ spec/features/header_handling/header_transformation_spec.rb
@@ -4,7 +4,7 @@ require 'spec_helper'
 
 fixture_path = 'spec/fixtures'
 
-describe 'be_able_to' do
+describe 'header transformations option' do
   it 'loads_file_with_dashes_in_header_fields as strings' do
     options = {strings_as_keys: true}
     data = SmarterCSV.process("#{fixture_path}/with_dashes.csv", options)
