+++ spec/features/header_handling/invalid_headers_spec.rb
