+++ spec/features/header_handling/keep_headers_spec.rb
@@ -4,17 +4,17 @@ require 'spec_helper'
 
 fixture_path = 'spec/fixtures'
 
-describe 'be_able_to' do
-  it 'strip_whitespace_from_headers' do
-    options = {strip_chars_from_headers: ' '}
+describe ':keep_original_headers option' do
+  it 'not_downcase_headers' do
+    options = {keep_original_headers: true}
     data = SmarterCSV.process("#{fixture_path}/basic.csv", options)
     expect(data.size).to eq 5
 
     data.each do |hash|
       hash.each_key do |key|
-        expect(key.class).to eq Symbol # all the keys should be symbols
+        expect(key.class).to eq String # all the keys should be string
 
-        expect(%i[firstname lastname dogs cats birds fish]).to include(key)
+        expect(['First Name', 'Last Name', 'Dogs', 'Cats', 'Birds', 'Fish']).to include(key)
       end
 
       expect(hash.size).to be <= 6
