+++ spec/features/header_handling/no_header_spec.rb
