+++ spec/features/header_handling/strings_as_keys_spec.rb
@@ -4,17 +4,17 @@ require 'spec_helper'
 
 fixture_path = 'spec/fixtures'
 
-describe 'be_able_to' do
-  it 'not_downcase_headers' do
-    options = {keep_original_headers: true}
+describe ':strings_as_keys option' do
+  it 'uses strings as hash keys' do
+    options = {strings_as_keys: true}
     data = SmarterCSV.process("#{fixture_path}/basic.csv", options)
     expect(data.size).to eq 5
 
     data.each do |hash|
       hash.each_key do |key|
-        expect(key.class).to eq String # all the keys should be string
+        expect(key.class).to eq String # all the keys should be symbols
 
-        expect(['First Name', 'Last Name', 'Dogs', 'Cats', 'Birds', 'Fish']).to include(key)
+        expect(%w[first_name last_name dogs cats birds fish]).to include(key)
       end
 
       expect(hash.size).to be <= 6
