+++ spec/features/special_cases/hard_sample_spec.rb
