+++ spec/features/special_cases/malformed_spec.rb
