+++ spec/features/special_cases/trading_spec.rb
